Switch to Client Side Rendering (#26)
The angular project was accidentally created with server-side rendering
which caused issues. This PR is to fix that. Also, this PR attempts to
make the deployment and docker builds as close as possible.

## Changes
- Remove all server files and configs
- Switch to psycopg2-binary 
- Create dockerfile for deployment
- Change nginx configuration
- Add environment variables for angular frontend
- Add DATABASE_PORT variable (this must now be set in the .env)
- Change API URL to be more consistent

## Testing
- Local docker build still works
- Deployment on the server works

## Outstanding Changes
- Need to ensure that only 1 entity can publish a certificate to the
blockchain but currently the publishing of certificates is linked to
each instance that is running this application. This can lead to a race
condition when using the same nonce or lead to failed transactions. Thus
we need to setup some way of handling this